refactor: deduplicate test suite, removing 14 redundant tests
- Delete test_token_bridge.py (5 tests), relocate 1 unique test to
  test_oauth_integration.py
- Remove TestTokenVerifierAudienceResolution from test_env_var_resolution.py
  (4 tests duplicated in test_token_verifier.py)
- Remove 3 duplicate DCR tests from test_oauth_metadata.py (covered by
  test_env_var_resolution.py)
- Remove TestAppLifespanOAuthMode from test_oauth_integration.py (3 tests
  duplicated in test_lifespan.py)
- Extract _default_claims() to shared tests/helpers.py
